requirements.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.35.0
django-environ==0.4.4
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.12.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.0.7
openapi-codec==1.3.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
